# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me/owned/{gameId}": {
      "post": {
        "operationId": "MeController_addOwned",
        "summary": "Add game to owned",
        "description": "Marks a game as owned by the current user.",
        "parameters": [
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "description": "Game UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Game added to owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameWrapperDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "MeController_removeOwned",
        "summary": "Remove game from owned",
        "description": "Removes the game from the current user’s owned list.",
        "parameters": [
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "description": "Game UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/me/owned": {
      "get": {
        "operationId": "MeController_listOwned",
        "summary": "List owned games",
        "description": "Returns the current user’s owned games, newest first.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Owned games",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/me/wishlist/{gameId}": {
      "post": {
        "operationId": "MeController_addWishlist",
        "summary": "Add game to wishlist",
        "description": "Marks a game as wishlist for the current user.",
        "parameters": [
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "description": "Game UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Game added to wishlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameWrapperDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "MeController_removeWishlist",
        "summary": "Remove game from wishlist",
        "description": "Removes the game from the current user’s wishlist.",
        "parameters": [
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "description": "Game UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/me/wishlist": {
      "get": {
        "operationId": "MeController_listWishlist",
        "summary": "List wishlist games",
        "description": "Returns the current user’s wishlist games, newest first.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Wishlist games",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/me/plays": {
      "post": {
        "operationId": "MeController_createPlayLog",
        "summary": "Create play log",
        "description": "Records a play of a game for the current user.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlayLogDto"
              },
              "examples": {
                "default": {
                  "summary": "Play entry",
                  "value": {
                    "gameId": "123e4567-e89b-12d3-a456-426614174000",
                    "playedAt": "2025-02-25T14:00:00.000Z",
                    "durationMinutes": 90,
                    "playersCount": 4,
                    "notes": "First play"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Play log created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayLogResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "MeController_listPlays",
        "summary": "List play logs",
        "description": "Returns the current user’s play logs, latest first (limit 50).",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Play logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayLogListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/me/plays/{id}": {
      "delete": {
        "operationId": "MeController_deletePlayLog",
        "summary": "Delete play log",
        "description": "Deletes a play log by ID. Only the owner can delete.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Play log UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Play log not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/me/follow/{username}": {
      "post": {
        "operationId": "MeController_follow",
        "summary": "Follow user",
        "description": "Follows the given username. Cannot follow yourself.",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "description": "Username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Following",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Cannot follow yourself",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "MeController_unfollow",
        "summary": "Unfollow user",
        "description": "Stops following the given username.",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "description": "Username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unfollowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Me"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "BG Hub API",
    "description": "Board game collection, wishlist, play logs and social features.",
    "version": "0.0.2",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GameWrapperDto": {
        "type": "object",
        "properties": {
          "game": {
            "description": "Game",
            "allOf": [
              {
                "$ref": "#/components/schemas/GameDto"
              }
            ]
          }
        },
        "required": [
          "game"
        ]
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Operation succeeded",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
      "GamesListResponseDto": {
        "type": "object",
        "properties": {
          "games": {
            "description": "List of games",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameDto"
            }
          }
        },
        "required": [
          "games"
        ]
      },
      "CreatePlayLogDto": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "string",
            "description": "UUID of the game played",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "playedAt": {
            "type": "string",
            "description": "When the game was played (ISO 8601)",
            "format": "date-time",
            "example": "2025-02-25T14:00:00.000Z"
          },
          "durationMinutes": {
            "type": "number",
            "description": "Duration in minutes",
            "minimum": 1
          },
          "playersCount": {
            "type": "number",
            "description": "Number of players",
            "minimum": 1
          },
          "notes": {
            "type": "string",
            "description": "Free-text notes"
          }
        },
        "required": [
          "gameId",
          "playedAt"
        ]
      },
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code",
            "example": 404
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Resource not found"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      }
    }
  }
}